+++ pkg/provider/server.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/cyberark/conjur-authn-k8s-client/pkg/log"
+	"github.com/cyberark/conjur-k8s-csi-provider/pkg/logmessages"
 	"google.golang.org/grpc"
 	"sigs.k8s.io/secrets-store-csi-driver/provider/v1alpha1"
 )
@@ -46,7 +47,7 @@ func newServerWithDeps(
 	mountFunc func(context.Context, *v1alpha1.MountRequest) (*v1alpha1.MountResponse, error),
 	versionFunc func(context.Context, *v1alpha1.VersionRequest) (*v1alpha1.VersionResponse, error),
 ) *ConjurProviderServer {
-	log.Info("Creating and registering gRPC server...")
+	log.Info(logmessages.CKCP018)
 	validateSocket(socketPath)
 
 	grpcServer := grpcFactory()
@@ -72,20 +73,21 @@ func (c *ConjurProviderServer) startWithDeps(
 	var err error
 	c.listener, err = listenerFactory("unix", socketPath)
 	if err != nil {
-		return fmt.Errorf("failed to start socket listener: %w", err)
+		log.Error(logmessages.CKCP020, err)
+		return fmt.Errorf(logmessages.CKCP020, err)
 	}
 
-	log.Info("Serving gRPC server on socket %s...", socketPath)
+	log.Info(logmessages.CKCP021, socketPath)
 	return c.grpcServer.Serve(c.listener)
 }
 
 // Stop halts the gRPC server and closes the socket listener.
 func (c *ConjurProviderServer) Stop() {
-	log.Info("Stopping gRPC server...")
+	log.Info(logmessages.CKCP022)
 
 	c.grpcServer.GracefulStop()
 
-	log.Info("gRPC server stopped.")
+	log.Info(logmessages.CKCP023)
 }
 
 func (c *ConjurProviderServer) Mount(ctx context.Context, req *v1alpha1.MountRequest) (*v1alpha1.MountResponse, error) {
@@ -101,9 +103,6 @@ func validateSocket(path string) {
 	if !strings.HasPrefix(dir, "/var/run/secrets-store-csi-providers") &&
 		!strings.HasPrefix(dir, "/etc/kubernetes/secrets-store-csi-providers") {
 
-		log.Warn("Using non-standard providers directory %s: "+
-			"Ensure this directory has been configured on your CSI Driver before proceeding",
-			dir,
-		)
+		log.Warn(logmessages.CKCP019, dir)
 	}
 }